refactor(prompts): 按 minigame 重整 prompt 模块结构（纯搬迁，零漂移）
把杂物筐 config/prompts/prompts_game.py 按「具体游戏 / 跨游戏共享 / 路由编排」三层拆开，
统一各小游戏的归类标准。此前 galgame、badminton 各有独立文件，soccer 却埋在 prompts_game.py，
badminton 还是 system/pregame 在 prompts_game.py、quick_lines 在 prompts_badminton.py 的半拆状态。

- 删除 prompts_game.py
- 新增 prompts_soccer.py：收 soccer 全部（system/quick_lines/pregame/anger_pressure 各 locale 变体 + 7 个 getter）
- prompts_badminton.py：并入此前留在 prompts_game.py 的 badminton system/pregame，收口半拆
- 新增 prompts_minigame_common.py：跨游戏共享 _normalize_prompt_lang / _localized_template /
  PREGAME_CONTEXT_INPUT_WATERMARK（含关键的 from prompts_sys import _loc）
- prompts_game_route.py -> prompts_minigame_route.py（git mv，内部符号不变，只改顶部 import 来源）
- galgame docstring 注明它是对话/剧情玩法、非 minigame、不共用 minigame_common（用 prompts_sys._loc）
- 改所有调用方 import（game_router + 4 个单测），无 re-export 门面

纯结构搬迁、整段搬运不重写字符串。已用 origin/main 旧版逐 getter 跨 16 locale x 7 mode
断言新模块输出与旧版字符全等（零漂移），相关 4 个单测全绿，docstring CJK 守门通过。
